made nav change when logged in/out
When you open up the website and are not logged in, it tells the user that they are not logged in and only have the login button showing.

When logged in with admin privilege, it says logged in and with what account (e.g. Admin),  and shows the logout button and the admin button which take you to the page to add and delete drivers. (No login button shows)

When logged in with user privilege, it says logged in and with the account (e.g. Redbull), and shows the logout page. It doesn't show the button to the admin page as they don't have an admin rank